# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DamascusTransit API",
    "description": "Real-time transit tracking and fleet management — multi-tenant SaaS edition.\n\n## Authentication\n\nMost read endpoints are public but require an `?operator=<slug>` query parameter.\n1. **POST /api/v1/auth/login** — exchange email/password for a JWT token\n2. Include the token as `Authorization: Bearer <token>` on protected endpoints\n\nRoles: `super_admin` · `admin` · `dispatcher` · `driver` · `viewer`\n\nTokens expire after 24 hours.",
    "contact": {
      "name": "Damascus Transit System",
      "email": "support@damascustransit.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/telemetry/json": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Telemetry frame (JSON form of telematics.proto VehicleStatus)",
        "description": "JSON variant — handy for dev + integration tests. In production the\nProtobuf endpoint is preferred for the 80% bandwidth saving.",
        "operationId": "telemetry_json_api_v1_telemetry_json_post",
        "parameters": [
          {
            "name": "x-device-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_PartialDecode"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/protobuf": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Telemetry frame (Protobuf VehicleStatus or VehicleStatusBatch)",
        "description": "Protobuf variant — production target.\n\nBody decoding is currently a placeholder: we accept any opaque payload,\nverify the HMAC, and return 204. Once the protoc-generated bindings\nland in api/proto/telematics_pb2.py, this handler parses the body,\nruns the same `_ingest` pipeline, and returns the same 204.",
        "operationId": "telemetry_protobuf_api_v1_telemetry_protobuf_post",
        "parameters": [
          {
            "name": "x-device-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Signature"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "_PartialDecode": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "speed_kph": {
            "type": "number",
            "title": "Speed Kph",
            "default": 0.0
          },
          "heading": {
            "type": "number",
            "title": "Heading",
            "default": 0.0
          },
          "engine_state": {
            "type": "boolean",
            "title": "Engine State",
            "default": false
          },
          "fuel_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Level"
          },
          "trigger_event": {
            "type": "integer",
            "title": "Trigger Event",
            "default": 0
          },
          "operator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Id"
          },
          "route_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id"
          },
          "driver_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "satellite_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satellite Count"
          },
          "hdop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hdop"
          },
          "cell_signal_dbm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell Signal Dbm"
          },
          "firmware_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Version"
          },
          "battery_mv": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Mv"
          },
          "is_replay": {
            "type": "boolean",
            "title": "Is Replay",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "timestamp",
          "latitude",
          "longitude"
        ],
        "title": "_PartialDecode",
        "description": "JSON shape that mirrors telematics.proto VehicleStatus 1:1.\nNames are snake_case to match the proto field names."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}